FA archiver: Fix enthusiastic compiler errors
Recent versions of gcc raised warnings if the return code from [v]asprintf
is not checked.  However, in the core of error handling this is painful to
report ... and in reality it's not going to fail, as the only likely cause
of failure is out of memory, in which case we're dead anyway!

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@52347 e099a375-04f9-0310-9d5f-a741eaff62e1