FA archiver: Make floating point error detection debug use only
Unfortunately there is a least one realistic case when normal operation
can generate a floating point exception, and the consequence is to halt
the archiver, so these exceptions are now only enabled when explicitly
selected.

In the case when monitoring the BPM timestamp as a position when the
timestamp makes a transition from 2^31-1 to -2^31 the decimation filter
can overshoot because of ringing, and then the conversion from double to
integer can generate an overflow exception.  Better to silently ignore
this than to crash!

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@61596 e099a375-04f9-0310-9d5f-a741eaff62e1