FA archiver: Use flock() instead of fcntl(F_SETLK)
Alas the quirks of fcntl(..., F_SETLK, ...) make the lock worthless: the
lock is dropped as soon as an archive capture is requested!  This occurs
because this class of lock is lost any time any file handle on the locked
file is closed.  Also, the fcntl lock is lost over the daemon() call in
archiver.c.  This should be fixed by using flock instead.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@61922 e099a375-04f9-0310-9d5f-a741eaff62e1