FA archiver: Rather substantial rework of buffer.c
Substantial rework and some simplification prompted by the following
thoughts: firstly, support for backlog reporting is currently unused,
and similarly there is only ever at most the one reserved reader, so
removing these options simplifies things.

Secondly, walking a list of readers on every write, particularly while
holding a lock, is just plain bad if it can be avoided ... and it so
happens it can.  Rewrite the detection of overrun to take place purely
in the context of the reader where it really belongs.

Also renamed stop_reader() to interrupt_reader() to be more honest,
particularly as it's inclined to busy-loop if used after being interrupted.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@79421 e099a375-04f9-0310-9d5f-a741eaff62e1