FA Archiver: Simplify implementation of reader
Reduce the number of major blocks returned as part of a DD read to 1.
This hugely simplifies the handling of indexes, there is now only one
index, and is a necessary preparatory step before we can start returning
detailed timestamps.

Unfortunately as currently written the number of socket writes increases
enormously.  We'll need to add a bit of buffering to avoid this problem.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@79426 e099a375-04f9-0310-9d5f-a741eaff62e1