FA archiver: Add support for reading timestamp from sniffer
If the sniffer device supports the new GET_TIMESTAMP ioctl use that
for timestamping sniffer blocks.  This reduces the jitter on individual
timestamps significantly as we can use a timestamp taken at the time of
interrupt.

Pick up latest snapshot of fa_sniffer.h from the driver which changes
some of the definitions to improve portability.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@79852 e099a375-04f9-0310-9d5f-a741eaff62e1