FA archiver: Add support for sending timestamps separately
To help with Java client support we'll want the option, particularly on
DD data, to read the timestamps at the end.  This ends up being a little
more complicated -- in large part because I don't want large scale
buffering to be under client control, so the server needs to take more care.

Also separate out pooled memory management into a separate file and
unify a number of different mechanisms.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@80437 e099a375-04f9-0310-9d5f-a741eaff62e1