FA archiver: Use madvise to preload DD pages
Because the double decimated area is memory mapped it can happen (on
startup or when the archiver rolls around) that the pages that need to
be updated are not mapped into memory.  Mapping these pages in can
take long enough to cause the archiver to drop data.

This commit fixes this by preloading every DD page at the start of the
cycle in which it will be used, giving 1.6 seconds notice.  Hopefully
this will be enough...

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@83380 e099a375-04f9-0310-9d5f-a741eaff62e1