FA archiver: Retrieve FA ids from server
Take advantage of the fact that the server now knows its name and knows
the list of available FA ids to retrieve the list of ids from the server
rather than from a file.

There is still a fallback for testing until the servers are update to
read from the fa-ids.sr file.  This will be deleted soon.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@114134 e099a375-04f9-0310-9d5f-a741eaff62e1