FA archiver: Move pthread broadcast notify inside locks
There five cases where pthread_cond_broadcast is used, and in all of them
the condition is signalled outside of the associated mutex.  This can be
more efficient (this avoids waking up a process when then immediately
blocks by taking the mutex which hasn't been released yet), but it's very
dangerous because a signal can end up being missed.

It's possible that we get away with this for legitimate reasons (no issues
have been seen), but it's *very* hazardous, so I'd rather revert to doing
this properly.

I suspect the reason we get away with this is that all of the uses here
are either from an events which recur continually, or which are very rare.

This is in effect a partial revert of SVN commit 79423, dated 2012/6/29.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@117259 e099a375-04f9-0310-9d5f-a741eaff62e1