FA archiver: Fix a subtle bug involving FA ids
The computation of the array of ids to retrieve involves some simple
integer arithmetic ... and it turns out that matlab has some unpleasant
surprises here.  In brief: int32(5)/10 == 1 .  Ah.

So, make sure we're working with doubles (fewer surprises here), both when
assembling the request mask in fa_load, and in the result returned by
fa_name2id.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@129797 e099a375-04f9-0310-9d5f-a741eaff62e1