FA archiver: A couple of signal handling fixes
At the ESRF it was found that when the archiver was launched by certain
applications the signal mask was blocking the shutdown signals, and so
the archiver was failing to terminate.  Fix this by setting the process
mask explicitly.

Also remove logging from the at_exit() call -- none of the logging code
is signal safe, so there's risk of nasty behaviour here.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@144403 e099a375-04f9-0310-9d5f-a741eaff62e1